Added tests to layoutmanageractionbeantest, moved functionality for updating linked components to the actionbean (listeners weren't fired in postgres/Stripersist.getEntityManager()).
